% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update_chart.R
\name{update_chart}
\alias{update_chart}
\title{Update Chart}
\usage{
update_chart(data, type, session)
}
\arguments{
\item{data}{A tibble containing the data to be sent to the chart.}

\item{type}{A string specifying which message handler or chart will receive the data.}

\item{session}{The Shiny session object used to send the custom message.}
}
\value{
na
}
\description{
Sends data as JSON in a custom message, which can be received by JavaScript and used to update an sszvis chart. This function requires an appropriate message handler in the JavaScript code specific to each chart type.
}
