% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statenttool-package.R
\docType{package}
\name{statenttool-package}
\alias{statenttool}
\alias{statenttool-package}
\title{statenttool: STATENT-Tool}
\description{
Ein Tool zum Zusammenstellen von Anzahl Arbeitsstätten, Beschäftigten und Betrieben nach Zeit, Raum, Branche, Rechtsform und Betriebsgrösse
}
\author{
\strong{Maintainer}: Aline Metzler \email{aline.metzler@zuerich.ch}

}
\keyword{internal}
