% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statenttool-package.R
\docType{package}
\name{statenttool-package}
\alias{statenttool}
\alias{statenttool-package}
\title{statenttool: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
